reviewed lecture 5, 6 and 7
Lecture 5: You use, and also really need, that the disturbance has the same distribution in any direction. Otherwise a uniform disturbance (on [-1,1]) in the x direction and an independent uniform disturbance (on [-1,1]) in the y direction would also satisfy, for example. 
Lecture 7: fixed some minor typos (and added one piece of additional explanation).